Added Layouts
Main window layout restoration and saving implemented.
