+++ gmk/src/gmkexception.cpp
@@ -21,7 +21,7 @@
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 **/
 
-#include <gmkexception.hpp>
+#include <gmkexception.h>
 
 namespace Gmk
 {
