+++ gmk/src/gmkgameinfo.cpp
@@ -21,7 +21,7 @@
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 **/
 
-#include <gmkgameinfo.hpp>
+#include <gmkgameinfo.h>
 
 namespace Gmk
 {
