+++ gmk/src/gmkincludefile.cpp
@@ -21,7 +21,7 @@
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 **/
 
-#include <gmkincludefile.hpp>
+#include <gmkincludefile.h>
 
 namespace Gmk
 {
