+++ gmk/src/gmkobject.cpp
@@ -21,8 +21,8 @@
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 **/
 
-#include <gmkobject.hpp>
-#include <gmk.hpp>
+#include <gmkobject.h>
+#include <gmk.h>
 
 namespace Gmk
 {
