+++ gmk/src/gmkrypt.cpp
@@ -21,7 +21,7 @@
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 **/
 
-#include <gmkrypt.hpp>
+#include <gmkrypt.h>
 
 namespace Gmk
 {
