+++ gmk/src/gmktrigger.cpp
@@ -21,7 +21,7 @@
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 **/
 
-#include <gmktrigger.hpp>
+#include <gmktrigger.h>
 
 namespace Gmk
 {
